refactor: Move serialization code out from `bitcoinconsensus` module
This change is a prerequisite for the following fix for
`libbitcoinconsensus-0.dll` builds with `DEBUG=1`.
