+++ src/Makefile.am
@@ -585,6 +585,8 @@ libbitcoin_consensus_a_SOURCES = \
   script/script.h \
   script/script_error.cpp \
   script/script_error.h \
+  script/transaction_unserialize.cpp \
+  script/transaction_unserialize.h \
   serialize.h \
   span.h \
   tinyformat.h \
